refactor: general clean up and org of all components
