+++ package.json
@@ -16,6 +16,7 @@
     "@blockstack/stacks-transactions": "^0.3.0-alpha.5",
     "@blockstack/ui": "^1.2.0-beta.9",
     "@reduxjs/toolkit": "^1.3.0",
+    "@styled-system/theme-get": "^5.1.2",
     "@types/prismjs": "^1.16.0",
     "babel-plugin-module-resolver": "^4.0.0",
     "blockstack": "^21.0.0",
@@ -26,7 +27,7 @@
     "fontfaceobserver": "^2.1.0",
     "formik": "^2.1.4",
     "isomorphic-unfetch": "^3.0.0",
-    "next": "^9.3.6-canary.11",
+    "next": "^9.3.6",
     "next-redux-wrapper": "^6.0.0-rc.7",
     "prism-react-renderer": "^1.1.0",
     "prismjs": "^1.20.0",
@@ -34,7 +35,6 @@
     "react-dom": "^16.13.1",
     "react-redux": "^7.2.0",
     "react-simple-code-editor": "^0.11.0",
-    "redaxios": "^0.2.0",
     "redux-thunk": "^2.3.0",
     "store": "^2.0.12",
     "styled-components": "^5.0.1",
@@ -55,6 +55,7 @@
     "@types/react-redux": "^7.1.7",
     "@types/redux-thunk": "^2.1.0",
     "@types/store": "^2.0.2",
+    "@types/styled-system__theme-get": "^5.0.0",
     "commitlint": "^8.3.5",
     "eslint": "^6.7.2",
     "eslint-plugin-prettier": "^3.1.3",
