+++ src/components/badge.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import { Flex, Text, FlexProps } from '@blockstack/ui';
+import { Flex, Text, BoxProps, FlexProps } from '@blockstack/ui';
 
-const Label: React.FC = props => (
+const Label: React.FC<BoxProps> = props => (
   <Text
     display="block"
     lineHeight="16px"
@@ -12,18 +12,16 @@ const Label: React.FC = props => (
   />
 );
 
-export const Badge: React.FC<FlexProps> = ({ children, ...rest }) => {
-  return (
-    <Flex
-      height="24px"
-      align="center"
-      justify="center"
-      borderRadius="24px"
-      py="extra-tight"
-      px="base"
-      {...rest}
-    >
-      <Label>{children}</Label>
-    </Flex>
-  );
-};
+export const Badge: React.FC<FlexProps> = ({ children, ...rest }) => (
+  <Flex
+    height="24px"
+    align="center"
+    justify="center"
+    borderRadius="24px"
+    py="extra-tight"
+    px="base"
+    {...rest}
+  >
+    <Label>{children}</Label>
+  </Flex>
+);
