+++ src/components/card.tsx
@@ -2,5 +2,11 @@ import * as React from 'react';
 import { Flex, FlexProps } from '@blockstack/ui';
 
 export const Card: React.FC<FlexProps> = props => (
-  <Flex flexDir="column" borderRadius="12px" border="1px solid" borderColor="var(--colors-border)" {...props} />
+  <Flex
+    flexDir="column"
+    borderRadius="12px"
+    border="1px solid"
+    borderColor="var(--colors-border)"
+    {...props}
+  />
 );
