+++ src/components/page.tsx
@@ -1,10 +1,8 @@
-import React, { useState } from 'react';
-import Link from 'next/link';
+import React from 'react';
 import { Box, BoxProps, Flex } from '@blockstack/ui';
+
+import { Header } from '@components/header';
 import { TransactionTitle, TitleProps } from '@components/transaction-title';
-import { BlockstackLogo } from '@components/icons/blockstack-logo';
-import { search } from '@common/search';
-import { SearchBarWithDropdown } from '@components/search-bar';
 import { FooterLinks } from '@components/footer-links';
 
 export const PageTop: React.FC<TitleProps> = ({ status, type, ...props }) => (
@@ -17,11 +15,11 @@ export const Page: React.FC<BoxProps> = ({ children, ...rest }) => (
   <Flex flexDirection="column" width="100%" minHeight="100%">
     <Flex
       as="main"
+      mx="auto"
+      width="100%"
+      flexGrow={1}
       height="100%"
       maxWidth="1280px"
-      mx="auto"
-      flexGrow={1}
-      width="100%"
       flexDirection="column"
       px={['base', 'extra-loose']}
       {...rest}
@@ -34,44 +32,9 @@ export const Page: React.FC<BoxProps> = ({ children, ...rest }) => (
   </Flex>
 );
 
-const Header = (props: any) => {
-  const [query, setQuery] = useState('');
-
-  const handleQueryUpdate = (e: React.ChangeEvent<HTMLInputElement>) => {
-    e.preventDefault();
-    setQuery(e.target.value);
-  };
-
-  const updateQuery = React.useCallback(handleQueryUpdate, []);
-  return (
-    <Flex
-      as="form"
-      onSubmit={search(query)}
-      height="64px"
-      flexDirection="row"
-      alignItems="center"
-      boxShadow="0px 1px 2px rgba(27, 39, 51, 0.04), 0px 4px 8px rgba(27, 39, 51, 0.04)"
-      borderBottom="1px solid"
-      borderColor="var(--colors-border)"
-      bg="var(--colors-bg)"
-      position="fixed"
-      width="100%"
-    >
-      <Link href="/" passHref>
-        <a aria-label="Homepage" title="Stacks Explorer Homepage">
-          <BlockstackLogo m="base-loose" />
-        </a>
-      </Link>
-      <SearchBarWithDropdown height="40px" maxWidth="544px" onChange={updateQuery} />
-    </Flex>
-  );
-};
-
-export const PageWrapper: React.FC<BoxProps> = props => {
-  return (
-    <>
-      <Header />
-      <Page pt="extra-loose" mt="64px" {...props} />
-    </>
-  );
-};
+export const PageWrapper: React.FC<BoxProps> = props => (
+  <>
+    <Header />
+    <Page pt="extra-loose" mt="64px" {...props} />
+  </>
+);
