+++ src/components/tx/coinbase.tsx
@@ -5,7 +5,6 @@ import { PageTop } from '@components/page';
 import { Rows } from '@components/rows';
 
 import { TransactionType } from '@models/transaction.interface';
-import { Statuses } from '@components/status';
 import { TransactionDetails } from '@components/transaction-details';
 import { CoinbaseTransaction } from '@blockstack/stacks-blockchain-sidecar-types';
 
@@ -13,28 +12,26 @@ interface CoinbasePageProps {
   transaction: CoinbaseTransaction;
 }
 
-const CoinbasePage = ({ transaction }: CoinbasePageProps) => {
-  return (
-    <>
-      <PageTop status={transaction.tx_status} type={[TransactionType.COINBASE]} />
-      <Stack spacing="extra-loose">
-        <TransactionDetails transaction={transaction} hideContract />
-        {transaction.coinbase_payload.data ? (
-          <Rows
-            noTopBorder
-            items={[
-              {
-                label: {
-                  children: 'Scratch space',
-                },
-                children: transaction.coinbase_payload.data,
+const CoinbasePage = ({ transaction }: CoinbasePageProps) => (
+  <>
+    <PageTop status={transaction.tx_status} type={[TransactionType.COINBASE]} />
+    <Stack spacing="extra-loose">
+      <TransactionDetails transaction={transaction} hideContract />
+      {transaction.coinbase_payload.data ? (
+        <Rows
+          noTopBorder
+          items={[
+            {
+              label: {
+                children: 'Scratch space',
               },
-            ]}
-          />
-        ) : null}
-      </Stack>
-    </>
-  );
-};
+              children: transaction.coinbase_payload.data,
+            },
+          ]}
+        />
+      ) : null}
+    </Stack>
+  </>
+);
 
 export default CoinbasePage;
