+++ src/components/tx/poison-microblock.tsx
@@ -12,31 +12,29 @@ interface PoisonMicroblockPageProps {
   transaction: PoisonMicroblockTransaction;
 }
 
-const PoisonMicroblockPage = ({ transaction }: PoisonMicroblockPageProps) => {
-  return (
-    <>
-      <PageTop status={transaction.tx_status} type={[TransactionType.POISON_MICROBLOCK]} />
-      <Stack spacing="extra-loose">
-        <TransactionDetails transaction={transaction} hideContract />
-        <Rows
-          items={[
-            {
-              label: {
-                children: 'Header 1',
-              },
-              children: '8a664c1dd163643b40a4e300daffdc1ea6af493d5ee7338ada9396b63b32dedc',
+const PoisonMicroblockPage = ({ transaction }: PoisonMicroblockPageProps) => (
+  <>
+    <PageTop status={transaction.tx_status} type={[TransactionType.POISON_MICROBLOCK]} />
+    <Stack spacing="extra-loose">
+      <TransactionDetails transaction={transaction} hideContract />
+      <Rows
+        items={[
+          {
+            label: {
+              children: 'Header 1',
             },
-            {
-              label: {
-                children: 'Header 2',
-              },
-              children: '8a664c1dd163643b40a4e300daffdc1ea6af493d5ee7338ada9396b63b32dedc',
+            children: '8a664c1dd163643b40a4e300daffdc1ea6af493d5ee7338ada9396b63b32dedc',
+          },
+          {
+            label: {
+              children: 'Header 2',
             },
-          ]}
-        />
-      </Stack>
-    </>
-  );
-};
+            children: '8a664c1dd163643b40a4e300daffdc1ea6af493d5ee7338ada9396b63b32dedc',
+          },
+        ]}
+      />
+    </Stack>
+  </>
+);
 
 export default PoisonMicroblockPage;
